# api-cms/docs/swagger.yaml
definitions:
  admin.Account:
    properties:
      link:
        type: string
      name:
        type: string
      platform:
        type: string
    type: object
  admin.Category:
    properties:
      name:
        type: string
    type: object
  admin.Keyword:
    properties:
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/v1/admin/accounts:
    get:
      consumes:
      - application/json
      description: Xem danh sách account
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xem danh sách account
      tags:
      - account
    post:
      consumes:
      - application/json
      description: Tạo mới account
      parameters:
      - description: name platform link
        in: body
        name: AccountForm
        required: true
        schema:
          $ref: '#/definitions/admin.Account'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Tạo mới account
      tags:
      - account
  /api/v1/admin/accounts/{account_id}:
    delete:
      consumes:
      - application/json
      description: Xóa account
      parameters:
      - description: account_id
        in: path
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xóa account
      tags:
      - account
    get:
      consumes:
      - application/json
      description: Xem chi tiết account
      parameters:
      - description: account_id
        in: path
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xem chi tiết account
      tags:
      - account
...
  /api/v1/common/categories:
    get:
      consumes:
      - application/json
      description: Xem danh sách Category
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xem danh sách Category
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: Tạo mới Category
      parameters:
      - description: Name
        in: body
        name: AccountForm
        required: true
        schema:
          $ref: '#/definitions/admin.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Tạo mới Category
      tags:
      - Category
  /api/v1/common/categories/{category_id}:
    delete:
      consumes:
      - application/json
      description: Xóa Category
      parameters:
      - description: category_id
        in: path
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xóa Category
      tags:
      - Category
    get:
      consumes:
      - application/json
      description: Xem chi tiết Category
      parameters:
      - description: category_id
        in: path
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xem chi tiết Category
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: Cập nhật Category
      parameters:
      - description: ID of the Category
        in: path
        name: category_id
        required: true
        type: string
      - description: Thông tin cập nhật của Category
        in: body
        name: CategoryForm
        required: true
        schema:
          $ref: '#/definitions/admin.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Cập nhật Category
      tags:
      - Category
  /api/v1/common/keywords:
    get:
      consumes:
      - application/json
      description: Xem danh sách Keyword
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xem danh sách Keyword
      tags:
      - Keyword
    post:
      consumes:
      - application/json
      description: Tạo mới Keyword
      parameters:
      - description: Name
        in: body
        name: AccountForm
        required: true
        schema:
          $ref: '#/definitions/admin.Keyword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Tạo mới Keyword
      tags:
      - Keyword
  /api/v1/common/keywords/{keyword_id}:
    delete:
      consumes:
      - application/json
      description: Xóa Keyword
      parameters:
      - description: Keyword_id
        in: path
        name: keyword_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xóa Keyword
      tags:
      - Keyword
    get:
      consumes:
      - application/json
      description: Xem chi tiết Keyword
      parameters:
      - description: Keyword_id
        in: path
        name: keyword_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xem chi tiết Keyword
      tags:
      - Keyword
    put:
      consumes:
      - application/json
      description: Sửa Keyword
      parameters:
      - description: Keyword_id
        in: path
        name: keyword_id
        required: true
        type: string
      - description: Thông tin cập nhật của Keyword
        in: body
        name: KeywordForm
        required: true
        schema:
          $ref: '#/definitions/admin.Keyword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Update Keyword
      tags:
      - Keyword
...
swagger: "2.0"